feat: migrate settings to standalone page and extract useOAuthBinding hook
