+++ web/src/route/main.tsx
@@ -1,6 +1,6 @@
+import LoadingPlaceholder from '@/components/others/LoadingPlaceholder';
 import ScrollPositionManager from '@/components/ScrollPositionManager';
 import LayoutDashboard from '@/layout/dashboard';
-import LoadingPlaceholder from '@/components/others/LoadingPlaceholder';
 import { useAuthState } from '@/state/profile';
 import { createBrowserRouter, Navigate, Outlet, RouterProvider } from 'react-router-dom';
 import { ProtectedRoute } from './protectedRoute';
@@ -19,6 +19,7 @@ import HomePage from '@/pages/home';
 import Landing from '@/pages/landing';
 import Login from '@/pages/login';
 import ProfilePage from '@/pages/profile';
+import SettingsPage from '@/pages/profile/setting';
 import SingleRotePage from '@/pages/rote/[roteid]';
 import SetupPage from '@/pages/setup';
 import UserPage from '@/pages/user/[username]';
@@ -131,6 +132,15 @@ export default function GlobalRouterProvider() {
               element: (
                 <ProtectedRoute>
                   <ProfilePage />
+                </ProtectedRoute>
+              ),
+              errorElement: <ErrorPage />,
+            },
+            {
+              path: 'profile/setting',
+              element: (
+                <ProtectedRoute>
+                  <SettingsPage />
                 </ProtectedRoute>
               ),
               errorElement: <ErrorPage />,
